refactor(admin): extract mount.ts and reduce block-form.ts to a re-export facade
mountBlockForm, its option/handle types and the wiring (value sync,
array add/delete/toggle, picker triggers, clear buttons, image-error
fallback, Sortable lifecycle) move verbatim into block-form/mount.ts.
block-form.ts is now an 18-line pure re-export facade over the
block-form/ modules — the ADR-0012 handlers.ts pattern — so both
editors and the dist consumers keep their import path unchanged.

The guard-test repointing planned for this slice landed in slice 5
(the moves broke them there); consumers verified untouched.

Part of #38, slice 6/6.
